Move check_partitioning_error() to common.py
The reason is to enable use of the function also
within details.py.
